+++ pages/vite.config.js
@@ -1,9 +1,0 @@
-import { getClientConfig } from '../viteConfigs'
-
-export default getClientConfig({
-  base: '',
-  root: '.',
-  build: {
-    outDir: '../pages_output',
-  },
-})